+++ src/main/java/com/example/end/controller/ReviewController.java
@@ -1,8 +1,10 @@
 package com.example.end.controller;
 
-import com.example.end.dto.ReviewDto;
 import com.example.end.models.Review;
 import com.example.end.service.interfaces.ReviewService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import com.example.end.dto.ReviewDto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
+
 
 @RestController
 @RequestMapping("/api/reviews")
@@ -29,7 +32,7 @@ public class ReviewController {
 
     @GetMapping("/{id}")
     @ApiOperation(value = "Get review by ID", response = Review.class)
-    public ResponseEntity<Review> getReviewById(@PathVariable int id) {
+    public ResponseEntity<Review> getReviewById(@PathVariable Long id) {
         Optional<Review> reviewOptional = reviewService.getReviewById(id);
         return reviewOptional.map(review -> ResponseEntity.ok().body(review))
                 .orElseGet(() -> ResponseEntity.notFound().build());
@@ -55,7 +58,7 @@ public class ReviewController {
 
     @PutMapping("/{id}")
     @ApiOperation(value = "Update existing review", response = Review.class)
-    public ResponseEntity<Review> updateReview(@PathVariable int id, @Valid @RequestBody ReviewDto reviewDto,
+    public ResponseEntity<Review> updateReview(@PathVariable Long id, @Valid @RequestBody ReviewDto reviewDto,
                                                BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return ResponseEntity.badRequest().build();
@@ -68,7 +71,7 @@ public class ReviewController {
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "Delete review by ID", response = HttpStatus.class)
-    public ResponseEntity<HttpStatus> deleteReview(@PathVariable int id) {
+    public ResponseEntity<HttpStatus> deleteReview(@PathVariable Long id) {
         if (reviewService.deleteReview(id)) {
             return ResponseEntity.ok().build();
         } else {
@@ -76,5 +79,4 @@ public class ReviewController {
         }
     }
 }
-
