+++ src/main/java/com/example/end/service/interfaces/ReviewService.java
@@ -6,17 +6,16 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-
 @Service
 public interface ReviewService {
     Review createReview(ReviewDto reviewDto);
 
     List<Review> getAllReviews();
 
-    Optional<Review> getReviewById(int id);
+    Optional<Review> getReviewById(Long id);
 
-    Optional<Review> updateReview(int id, ReviewDto reviewDto);
+    Optional<Review> updateReview(Long id, ReviewDto reviewDto);
 
-    boolean deleteReview(int id);
+    boolean deleteReview(Long id);
     // Методы для управления отзывами
 }
